refactor(helix-tui): split `buffer.rs` into submodules
